% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{KMP}
\alias{KMP}
\title{Gives vector of positons in text S in which P is found}
\usage{
KMP(P, S)
}
\arguments{
\item{P}{The substring/pattern to search in text S}

\item{S}{The text in which P will be searched}
}
\value{
A vector of integers, that will represent the positions
in S where a complete match is found. These positions represent the
position of the first character of P in S.If no occurence is found NULL is given.
}
\description{
This function compute a vector representing positions at which a substring P is present
in a text S. if present, it recognizes spaces as part of the substring P. The search is
done through the Knuth-Morris-Pratt algorithm
}
\references{
\url{https://en.wikipedia.org/wiki/Knuth-Morris-Pratt_algorithm}\cr
}
\seealso{
\code{\link{kmp_index}}\cr
}
\author{
Lorenzo Brochier\cr e-mail: \href{mailto:lorenzomathieu.brochier@mail.polimi.it}{lorenzomathieu.brochier@mail.polimi.it}
}
