% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{kmp_index}
\alias{kmp_index}
\title{Compute the index of a word, using Knuth-Morris-Pratt algorithm}
\usage{
kmp_index(pattern)
}
\arguments{
\item{pattern}{The word for which the index will be created}
}
\value{
A vector of integers, that will represent the index for the given word.
}
\description{
This function compute the index for the KMP algorithm. Inside the function the index
will be 1-based, but the output will be 0-based as described in the wikipedia page
}
\references{
\url{https://en.wikipedia.org/wiki/Knuth-Morris-Pratt_algorithm}\cr
}
\seealso{
\code{\link{KMP}}\cr
}
\author{
Lorenzo Brochier\cr e-mail: \href{mailto:lorenzomathieu.brochier@mail.polimi.it}{lorenzomathieu.brochier@mail.polimi.it}
}
